Change epub main class name and remove epub3 class
